requirements.txt
cython
scipy
xmltodict
tqdm
transformers==4.5.1
onnx
onnxruntime
prettytable
submitit
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
timm
deprecated
django
easydict
ete3
future
ipython
ipdb
jinja2
matplotlib
nltk
numpy
opencv-python
orderedset
pandas
pathos
pillow
progressbar
protobuf
psutil
pymongo
python-magic
pyyaml
ruamel.yaml
seaborn
simplejson
sklearn
scikit-image
traceback2
tb-nightly
yacs
torchlars
boto3
anytree
botocore
einops